parser/pageparser: File renames and splitting
See #5324
